Fix overflow
