+++ lib/experimental/PageLayouts/Utils/DetailsItemsList/index.tsx
