+++ lib/experimental/PageLayouts/Utils/exports.tsx
@@ -1,4 +1,3 @@
 export * from "./DetailsItem"
 export * from "./DetailsItemsList"
 export * from "./TagsList"
-export * from "./VerticalSeparator"
