+++ lib/experimental/exports.ts
@@ -3,9 +3,9 @@ import { ScrollArea as ScrollAreaComponent } from "./Utilities/ScrollArea"
 
 export * from "./Forms/exports"
 export * from "./Information/exports"
-export * from "./Layouts/exports"
 export * from "./Navigation/exports"
 export * from "./Overlays/exports"
+export * from "./PageLayouts/exports"
 export * from "./Utilities/exports"
 export * from "./Widgets/exports"
